Reworked XML Serialization
Updated the Timeline data storage from a list variable to a class, making it easier to serialize and access. Moved the timeline onto the main form and added code sectioning the timeline according to the amount of events and locations loaded. Added ability to mark people as present for an event. Reworked the UI to be easier to use.